Crate clfft

This module provides access to the clFFT lib in a fashion which is more consistent with the ocl crate. Users can always use the ffi module to directly access clFFT.

Keep in mind that this crate just provides the bindings to clFFT. In order to build it the linker needs to have clFFT.lib available and in order to run it the clFFT library needs to be there. See the build section in the README for further details.

Modules

ffi
Bindings to clFFT. See http://clmathlibraries.github.io/clFFT/ for more information.

Structs

FftInplacePlan
A plan is a repository of state for calculating FFT's. Allows the runtime to pre-calculate kernels, programs and buffers and associate them with buffers of specified dimensions.

FftOutOfPlacePlan
A plan is a repository of state for calculating FFT's. Allows the runtime to pre-calculate kernels, programs and buffers and associate them with buffers of specified dimensions.

FftPlanBuilder
Builder for a FFT plan.

Enums

Direction
Specify the expected direction of each FFT, time or the frequency domains

Layout
Specify the expected layouts of the buffers.

Location
pecify wheter the input buffers are overwritten with results

Precision
Specify the expected precision of each FFT.

Traits

AsClFftPlanHandle
Gets the native clFFT plan handle from a type.

ClFftPrm
A trait for all paremeters supported by clFFT.

FftPlan
Getters for a FFT plan.

Functions

builder
Creates a builder for baking a new FFT plan.

teatdown
Frees all clFFT library resources. After calling this function no further methods must be called on this lib!