Struct clfft::FftOutOfPlacePlan

pub struct FftOutOfPlacePlan<'a, T: ClFftPrm> { /* fields omitted */ }

A plan is a repository of state for calculating FFT's. Allows the runtime to pre-calculate kernels, programs and buffers and associate them with buffers of specified dimensions.

Methods

impl<'a, T: ClFftPrm> FftOutOfPlacePlan<'a, T>

fn ewait(self, wait_list: &'a ClWaitList) -> Self

Specifies the list of events to wait on before the command will run.

fn ewait_opt(self, wait_list: Option<&'a ClWaitList>) -> Self

Specifies a list of events to wait on before the command will run.

fn enew(self, new_event_dest: &'a mut ClEventPtrNew) -> Self

Specifies the destination list or empty event for a new, optionally created event associated with this command.

fn enew_opt(self, new_event_list: Option<&'a mut ClEventPtrNew>) -> Self

Specifies a destination list for a new, optionally created event associated with this command.

fn enq(&mut self,
direction: Direction,
buffer: &Buffer<T>,
result: &mut Buffer<T>)
-> Result<()>

Enqueues the FFT so that it gets performed on the device.