Struct clfft::FftPlanBuilder

pub struct FftPlanBuilder<T: ClFftPrm> { /* fields omitted */ }

Builder for a FFT plan.

Methods

impl<T: ClFftPrm> FftPlanBuilder<T>

fn precision(self, precision: Precision) -> Self

Set the floating point precision of the FFT data.

fn input_layout(self, input_layout: Layout) -> Self

Set the expected layout of the input buffer.

fn output_layout(self, output_layout: Layout) -> Self

Set the expected layout of the output buffer.

fn dims<D: Into<SpatialDims>>(self, dims: D) -> Self

Set the dimensionality of the FFT transform; describes how many elements are in the array.

If the data is complex then the dimesions are specified are per complex number. In practice that means that the dimensions should be half the size of the buffers.

fn forward_scale(self, scale: f32) -> Self

Set the scaling factor that is applied to the FFT data.

fn backward_scale(self, scale: f32) -> Self

Set the scaling factor that is applied to the FFT data.

fn batch_size(self, scale: usize) -> Self

Set the number of discrete arrays that the plan can concurrently handle.

fn bake_inplace_plan<'a>(&mut self,
pro_que: &'a ProQue)
-> Result<FftInplacePlan<'a, T>>

Creates a plan for an inplace FFT.

fn bake_out_of_place_plan<'a>(&mut self,
pro_que: &'a ProQue)
-> Result<FftOutOfPlacePlan<'a, T>>

Creates a plan for an out of place FFT.